fix: fixed linting issues, refactoring some code along the way
affects: @postdfm/dfm2ast
