+++ packages/@postdfm/dfm2ast/src/ast/astNode.ts
@@ -1,0 +1,8 @@
+import { ASTType } from "./astType";
+
+export class ASTNode {
+  public astType: ASTType;
+  constructor(astType: ASTType) {
+    this.astType = astType;
+  }
+}
