+++ packages/@postdfm/dfm2ast/src/ast/astType.ts
@@ -1,0 +1,14 @@
+export enum ASTType {
+  String = "string",
+  HexString = "hexString",
+  Integer = "integer",
+  Double = "double",
+  Boolean = "boolean",
+  Qualified = "qualified",
+  Item = "item",
+  StringList = "stringList",
+  QualifiedList = "qualifiedList",
+  ItemList = "itemList",
+  Property = "property",
+  Object = "object"
+}
