+++ packages/@postdfm/dfm2ast/src/ast/list/itemList.ts
@@ -1,0 +1,9 @@
+import { ASTType } from "../astType";
+import { Item } from "../item";
+import { List } from "./list";
+
+export class ItemList extends List<Item> {
+  constructor(values: Item[]) {
+    super(ASTType.ItemList, values);
+  }
+}
