+++ packages/@postdfm/dfm2ast/src/ast/list/qualifiedList.ts
@@ -1,0 +1,9 @@
+import { ASTType } from "../astType";
+import { QualifiedValue } from "../value/qualifiedValue";
+import { List } from "./list";
+
+export class QualifiedList extends List<QualifiedValue> {
+  constructor(values: QualifiedValue[]) {
+    super(ASTType.QualifiedList, values);
+  }
+}
