+++ packages/@postdfm/dfm2ast/src/ast/list/stringList.ts
@@ -1,0 +1,9 @@
+import { ASTType } from "../astType";
+import { StringValue } from "../value/stringValue";
+import { List } from "./list";
+
+export class StringList extends List<StringValue> {
+  constructor(values: StringValue[]) {
+    super(ASTType.StringList, values);
+  }
+}
