+++ packages/@postdfm/dfm2ast/src/ast/objectKind.ts
@@ -1,0 +1,5 @@
+export enum ObjectKind {
+  Object = "object",
+  Inline = "inline",
+  Inherited = "inherited"
+}
