+++ packages/@postdfm/dfm2ast/src/ast/value/booleanValue.ts
@@ -1,0 +1,8 @@
+import { ASTType } from "../astType";
+import { Value } from "./value";
+
+export class BooleanValue extends Value<boolean> {
+  constructor(value: boolean) {
+    super(ASTType.Boolean, value);
+  }
+}
