+++ packages/@postdfm/dfm2ast/src/ast/value/doubleValue.ts
@@ -1,0 +1,8 @@
+import { ASTType } from "../astType";
+import { Value } from "./value";
+
+export class DoubleValue extends Value<number> {
+  constructor(value: number) {
+    super(ASTType.Double, value);
+  }
+}
