+++ packages/@postdfm/dfm2ast/src/ast/value/hexStringValue.ts
@@ -1,0 +1,8 @@
+import { ASTType } from "../astType";
+import { Value } from "./value";
+
+export class HexStringValue extends Value<string> {
+  constructor(value: string) {
+    super(ASTType.HexString, value);
+  }
+}
