+++ packages/@postdfm/dfm2ast/src/ast/value/integerValue.ts
@@ -1,0 +1,8 @@
+import { ASTType } from "../astType";
+import { Value } from "./value";
+
+export class IntegerValue extends Value<number> {
+  constructor(value: number) {
+    super(ASTType.Integer, value);
+  }
+}
