+++ packages/@postdfm/dfm2ast/src/ast/value/qualifiedValue.ts
@@ -1,0 +1,8 @@
+import { ASTType } from "../astType";
+import { Value } from "./value";
+
+export class QualifiedValue extends Value<string> {
+  constructor(value: string) {
+    super(ASTType.Qualified, value);
+  }
+}
