+++ packages/@postdfm/dfm2ast/src/ast/value/stringValue.ts
@@ -1,0 +1,8 @@
+import { ASTType } from "../astType";
+import { Value } from "./value";
+
+export class StringValue extends Value<string> {
+  constructor(value: string) {
+    super(ASTType.String, value);
+  }
+}
