+++ packages/@postdfm/dfm2ast/src/ast/value/value.ts
@@ -1,0 +1,10 @@
+import { ASTNode } from "../astNode";
+import { ASTType } from "../astType";
+
+export class Value<T> extends ASTNode {
+  public value: T;
+  constructor(astType: ASTType, value: T) {
+    super(astType);
+    this.value = value;
+  }
+}
